pyproject.toml
[project]
name = "aws-identity-center-config-file-generator"
version = "0.1.0"
description = "Generates config files for Identity Center users for use by the AWS CLI/SDKs "
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.42.78",
    "inquirer>=3.4.1",
    "jinja2>=3.1.6",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "ruff>=0.15.8",
]
